Rename `rten_text::{tokenizers => tokenizer, normalizer => normalizers}`
 - The `tokenizer` module no longer contains models in sub-modules, but
   just the `Tokenizer` struct that runs the whole pipeline

 - Pluralize `normalizers` to match the `models` and `pretokenizers`
   modules
