+++ rten-examples/src/distilvit.rs
@@ -8,7 +8,7 @@ use rten_generate::{Generator, GeneratorUtils};
 use rten_imageio::read_image;
 use rten_tensor::prelude::*;
 use rten_tensor::NdTensor;
-use rten_text::tokenizers::Tokenizer;
+use rten_text::tokenizer::Tokenizer;
 
 struct Args {
     encoder_model: String,
