+++ rten-examples/src/trocr.rs
@@ -9,7 +9,7 @@ use rten_imageio::read_image;
 use rten_imageproc::normalize_image;
 use rten_tensor::prelude::*;
 use rten_tensor::NdTensor;
-use rten_text::tokenizers::Tokenizer;
+use rten_text::tokenizer::Tokenizer;
 
 struct Args {
     encoder_model: String,
