+++ rten-text/src/lib.rs
@@ -9,8 +9,8 @@
 //! [HuggingFace tokenizers](https://github.com/huggingface/tokenizers).
 
 pub mod models;
-pub mod normalizer;
+pub mod normalizers;
 pub mod pre_tokenizers;
-pub mod tokenizers;
+pub mod tokenizer;
 
 mod split;
